study(step1): 코틀린 DSL 생성한 클래스 패키지 이동 및 정리
