+++ src/main/kotlin/study/Introduce.kt
@@ -1,0 +1,5 @@
+package study
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    return PersonBuilder().apply(block).build()
+}
