+++ src/main/kotlin/study/Person.kt
@@ -1,0 +1,3 @@
+package study
+
+data class Person(val name: String, val company: String, val skills: Skill, val languages: List<Language>)
