+++ src/main/kotlin/study/Skill.kt
@@ -1,0 +1,3 @@
+package study
+
+data class Skill(val softSkills: List<String>, val hardSkills: List<String>)
