+++ src/main/kotlin/study/SkillBuilder.kt
@@ -1,0 +1,18 @@
+package study
+
+class SkillBuilder {
+    private var softSkills: MutableList<String> = mutableListOf()
+    private var hardSkills: MutableList<String> = mutableListOf()
+
+    fun soft(value: String) {
+        softSkills.add(value)
+    }
+
+    fun hard(value: String) {
+        hardSkills.add(value)
+    }
+
+    fun build(): Skill {
+        return Skill(softSkills.toList(), hardSkills.toList())
+    }
+}
